fix: constrain font subtype to one of allowed values
