+++ pdfgen/src/document/mod.rs
@@ -1,7 +1,11 @@
 use std::io::{Error, Write};
 
 use crate::types::{
-    hierarchy::{catalog::Catalog, page_tree::PageTree, primitives::font::Font},
+    hierarchy::{
+        catalog::Catalog,
+        page_tree::PageTree,
+        primitives::font::{Font, FontSubtype},
+    },
     page::Page,
     pdf_writer::PdfWriter,
 };
@@ -71,7 +75,7 @@ impl Document {
     }
 
     /// Creates a new font inside the document.
-    pub fn create_font(&mut self, subtype: Vec<u8>, base_type: Vec<u8>) -> ObjId<Font> {
+    pub fn create_font(&mut self, subtype: FontSubtype, base_type: Vec<u8>) -> ObjId<Font> {
         let id = self.id_manager.create_id();
 
         self.fonts.push(Font::new(id.clone(), subtype, base_type));
@@ -119,12 +123,15 @@ impl Document {
 
 #[cfg(test)]
 mod tests {
-    use crate::{Document, types::hierarchy::primitives::rectangle::Rectangle};
+    use crate::{
+        Document,
+        types::hierarchy::primitives::{font::FontSubtype, rectangle::Rectangle},
+    };
 
     fn create_sample_doc() -> Document {
         let mut document = Document::default();
         document.create_page().set_mediabox(Rectangle::A4);
-        document.create_font("Type1".into(), "Helvetica".into());
+        document.create_font(FontSubtype::Type1, "Helvetica".into());
 
         document
     }
